# [Cache] Allow to use namespace delimiter in cache key