--- a/src/Symfony/Component/Cache/Adapter/AbstractTagAwareAdapter.php
+++ b/src/Symfony/Component/Cache/Adapter/AbstractTagAwareAdapter.php
@@ -49,7 +49,7 @@ abstract class AbstractTagAwareAdapter implements TagAwareAdapterInterface, TagA
 
     protected function __construct(string $namespace = '', int $defaultLifetime = 0)
     {
-        if (static::$reservedChars === null) {
+        if (null === static::$reservedChars) {
             static::$reservedChars = str_replace(self::NS_SEPARATOR, '', CacheItem::RESERVED_CHARACTERS);
         }
         $this->namespace = '' === $namespace ? '' : CacheItem::validateKey($namespace, static::$reservedChars).static::NS_SEPARATOR;