--- a/src/Symfony/Component/Cache/Adapter/ArrayAdapter.php
+++ b/src/Symfony/Component/Cache/Adapter/ArrayAdapter.php
@@ -35,7 +35,11 @@ class ArrayAdapter implements AdapterInterface, CacheInterface, LoggerAwareInter
      */
     protected const NS_SEPARATOR = ':';
 
-    private bool $storeSerialized;
+    /**
+     * @internal
+     */
+    protected static ?string $reservedChars = null;
+
     private array $values = [];
     private array $tags = [];
     private array $expiries = [];
@@ -51,6 +55,10 @@ class ArrayAdapter implements AdapterInterface, CacheInterface, LoggerAwareInter
         private float $maxLifetime = 0,
         private int $maxItems = 0,
     ) {
+        if (static::$reservedChars === null) {
+            static::$reservedChars = str_replace(self::NS_SEPARATOR, '', CacheItem::RESERVED_CHARACTERS);
+        }
+
         if (0 > $maxLifetime) {
             throw new InvalidArgumentException(sprintf('Argument $maxLifetime must be positive, %F passed.', $maxLifetime));
         }
@@ -110,7 +118,7 @@ class ArrayAdapter implements AdapterInterface, CacheInterface, LoggerAwareInter
 
             return true;
         }
-        \assert('' !== CacheItem::validateKey($key, static::NS_SEPARATOR));
+        \assert('' !== CacheItem::validateKey($key, static::$reservedChars));
 
         return isset($this->expiries[$key]) && !$this->deleteItem($key);
     }
@@ -140,7 +148,7 @@ class ArrayAdapter implements AdapterInterface, CacheInterface, LoggerAwareInter
 
     public function deleteItem(mixed $key): bool
     {
-        \assert('' !== CacheItem::validateKey($key, static::NS_SEPARATOR));
+        \assert('' !== CacheItem::validateKey($key, static::$reservedChars));
         unset($this->values[$key], $this->tags[$key], $this->expiries[$key]);
 
         return true;
@@ -356,7 +364,7 @@ class ArrayAdapter implements AdapterInterface, CacheInterface, LoggerAwareInter
     {
         foreach ($keys as $key) {
             if (!\is_string($key) || !isset($this->expiries[$key])) {
-                CacheItem::validateKey($key, static::NS_SEPARATOR);
+                CacheItem::validateKey($key, static::$reservedChars);
             }
         }