--- a/src/Symfony/Component/Cache/CacheItem.php
+++ b/src/Symfony/Component/Cache/CacheItem.php
@@ -124,10 +124,11 @@ final class CacheItem implements ItemInterface
      * Validates a cache key according to PSR-6.
      *
      * @param mixed $key The key to validate
+     * @param string $reservedChars Can be used to override the list of reserved characters
      *
      * @throws InvalidArgumentException When $key is not valid
      */
-    public static function validateKey($key, string $allowChars = null): string
+    public static function validateKey($key, string $reservedChars = self::RESERVED_CHARACTERS): string
     {
         if (!\is_string($key)) {
             throw new InvalidArgumentException(sprintf('Cache key must be string, "%s" given.', get_debug_type($key)));
@@ -135,7 +136,6 @@ final class CacheItem implements ItemInterface
         if ('' === $key) {
             throw new InvalidArgumentException('Cache key length must be greater than zero.');
         }
-        $reservedChars = null === $allowChars ? self::RESERVED_CHARACTERS : str_replace(str_split($allowChars), '', self::RESERVED_CHARACTERS);
         if ('' !== $reservedChars && false !== strpbrk($key, $reservedChars)) {
             throw new InvalidArgumentException(sprintf('Cache key "%s" contains reserved characters "%s".', $key, $reservedChars));
         }