--- a/src/Symfony/Component/Cache/Tests/Adapter/Psr16AdapterTest.php
+++ b/src/Symfony/Component/Cache/Tests/Adapter/Psr16AdapterTest.php
@@ -12,9 +12,11 @@
 namespace Symfony\Component\Cache\Tests\Adapter;
 
 use Psr\Cache\CacheItemPoolInterface;
+use Psr\Log\AbstractLogger;
 use Symfony\Component\Cache\Adapter\ArrayAdapter;
 use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 use Symfony\Component\Cache\Adapter\Psr16Adapter;
+use Symfony\Component\Cache\Exception\InvalidArgumentException;
 use Symfony\Component\Cache\Psr16Cache;
 
 /**
@@ -29,9 +31,105 @@ class Psr16AdapterTest extends AdapterTestCase
         'testClearPrefix' => 'SimpleCache cannot clear by prefix',
     ];
 
+    private TestLogger $testLogger;
+
     public function createCachePool(int $defaultLifetime = 0): CacheItemPoolInterface
     {
-        return new Psr16Adapter(new Psr16Cache(new FilesystemAdapter()), '', $defaultLifetime);
+        $this->testLogger = new TestLogger();
+        $psr16Adapter = new Psr16Adapter(new Psr16Cache(new FilesystemAdapter()), '', $defaultLifetime);
+        $psr16Adapter->setLogger($this->testLogger);
+
+        return $psr16Adapter;
+    }
+
+    /**
+     * @dataProvider invalidKeys
+     */
+    public function testGetItemInvalidKeys($key)
+    {
+        if (isset($this->skippedTests[__FUNCTION__])) {
+            $this->markTestSkipped($this->skippedTests[__FUNCTION__]);
+        }
+
+        try {
+            $this->cache->getItem($key);
+        } catch (\Exception $exception) {
+            $this->assertInstanceOf(InvalidArgumentException::class, $exception);
+
+            return;
+        }
+
+        $this->assertNotEmpty($this->testLogger->records);
+        $record = $this->testLogger->records[0];
+
+        $this->assertSame('warning', $record['level']);
+        $this->assertSame(sprintf('Failed to fetch key "{key}": Cache key "%s" contains reserved characters "{}()/\\@:".', $key), $record['message']);
+        $this->assertSame('Symfony\\Component\\Cache\\Adapter\\Psr16Adapter', $record['context']['cache-adapter']);
+        $this->assertSame($key, $record['context']['key']);
+
+        $exception = $record['context']['exception'];
+        $this->assertInstanceOf(InvalidArgumentException::class, $exception);
+        $this->assertSame(sprintf('Cache key "%s" contains reserved characters "{}()/\\@:".', $key), $exception->getMessage());
+    }
+
+    /**
+     * @dataProvider invalidKeys
+     */
+    public function testHasItemInvalidKeys($key)
+    {
+        if (isset($this->skippedTests[__FUNCTION__])) {
+            $this->markTestSkipped($this->skippedTests[__FUNCTION__]);
+        }
+
+        try {
+            $this->cache->hasItem($key);
+        } catch (\Exception $exception) {
+            $this->assertInstanceOf(InvalidArgumentException::class, $exception);
+
+            return;
+        }
+
+        $this->assertNotEmpty($this->testLogger->records);
+        $record = $this->testLogger->records[0];
+
+        $this->assertSame('warning', $record['level']);
+        $this->assertSame(sprintf('Failed to check if key "{key}" is cached: Cache key "%s" contains reserved characters "{}()/\@:".', $key), $record['message']);
+        $this->assertSame('Symfony\\Component\\Cache\\Adapter\\Psr16Adapter', $record['context']['cache-adapter']);
+        $this->assertSame($key, $record['context']['key']);
+
+        $exception = $record['context']['exception'];
+        $this->assertInstanceOf(InvalidArgumentException::class, $exception);
+        $this->assertSame(sprintf('Cache key "%s" contains reserved characters "{}()/\\@:".', $key), $exception->getMessage());
+    }
+
+    /**
+     * @dataProvider invalidKeys
+     */
+    public function testDeleteItemInvalidKeys($key)
+    {
+        if (isset($this->skippedTests[__FUNCTION__])) {
+            $this->markTestSkipped($this->skippedTests[__FUNCTION__]);
+        }
+
+        try {
+            $this->cache->deleteItem($key);
+        } catch (\Exception $exception) {
+            $this->assertInstanceOf(InvalidArgumentException::class, $exception);
+
+            return;
+        }
+
+        $this->assertNotEmpty($this->testLogger->records);
+        $record = $this->testLogger->records[0];
+
+        $this->assertSame('warning', $record['level']);
+        $this->assertSame(sprintf('Failed to delete key "{key}": Cache key "%s" contains reserved characters "{}()/\@:".', $key), $record['message']);
+        $this->assertSame('Symfony\\Component\\Cache\\Adapter\\Psr16Adapter', $record['context']['cache-adapter']);
+        $this->assertSame($key, $record['context']['key']);
+
+        $exception = $record['context']['exception'];
+        $this->assertInstanceOf(InvalidArgumentException::class, $exception);
+        $this->assertSame(sprintf('Cache key "%s" contains reserved characters "{}()/\\@:".', $key), $exception->getMessage());
     }
 
     public function testValidCacheKeyWithNamespace()
@@ -42,5 +140,19 @@ class Psr16AdapterTest extends AdapterTestCase
         $cache->save($item);
 
         $this->assertTrue($cache->getItem('my_key')->isHit(), 'Stored item is successfully retrieved.');
+    }
+}
+
+final class TestLogger extends AbstractLogger
+{
+    public array $records = [];
+
+    public function log($level, $message, array $context = []): void
+    {
+        $this->records[] = [
+            'level' => $level,
+            'message' => $message,
+            'context' => $context,
+        ];
     }
 }