--- a/src/Symfony/Component/Cache/Traits/AbstractAdapterTrait.php
+++ b/src/Symfony/Component/Cache/Traits/AbstractAdapterTrait.php
@@ -343,7 +343,7 @@ trait AbstractAdapterTrait
         if (\is_string($key) && isset($this->ids[$key])) {
             return $this->namespace.$this->namespaceVersion.$this->ids[$key];
         }
-        \assert('' !== CacheItem::validateKey($key, static::NS_SEPARATOR));
+        \assert('' !== CacheItem::validateKey($key, static::$reservedChars));
         $this->ids[$key] = $key;
 
         if (\count($this->ids) > 1000) {