| Q             | A
| ------------- | ---
| Branch?       | 7.2
| Bug fix?      | no
| New feature?  | yes
| Deprecations? | no
| Issues        | https://github.com/symfony/symfony/issues/45599
| License       | MIT

Replaces https://github.com/symfony/symfony/pull/51603

This PR allow colon char `:` in cache key. It may be useful for redis grouping keys by pattern without creating a many pools for each namespace.

Difference between implementation https://github.com/symfony/symfony/pull/47561

- colon is only allowed for Symfony contracts cache, the PSR-6/16 adapters keeps this validation as before.
- Allow colon in namespace name too

--------------------------------------------

@nicolas-grekas I have taken a look at some approaches to how to optimize this, but it looks like:
```php
$reservedChars = null === $allowChars 
    ? self::RESERVED_CHARACTERS
    : str_replace(str_split($allowChars), '', self::RESERVED_CHARACTERS);
```
is pretty well optimized and is marginally faster than `preg_replace` and a lot faster than some other `for` based approaches I tried. I think the main problem is that this code would need to get executed for each validation, of which there can be many in single http request. So I thought that maybe a different approach would could work - simply allowing override of RESERVED_CHARACTERS via the parameter. Here is a simple test for performance comparison.

```php
<?php

class CacheItem {
    private const RESERVED_CHARACTERS = '{}()/\@:';

    /**
     * Validates a cache key according to PSR-6.
     *
     * @param mixed $key The key to validate
     *
     * @throws InvalidArgumentException When $key is not valid
     */
    public static function validateKey($key, string $allowChars = null): string
    {
        if (!\is_string($key)) {
            throw new InvalidArgumentException(sprintf('Cache key must be string, "%s" given.', get_debug_type($key)));
        }
        if ('' === $key) {
            throw new InvalidArgumentException('Cache key length must be greater than zero.');
        }
        $reservedChars = null === $allowChars ? self::RESERVED_CHARACTERS : str_replace(str_split($allowChars), '', self::RESERVED_CHARACTERS);
        if ('' !== $reservedChars && false !== strpbrk($key, $reservedChars)) {
            throw new InvalidArgumentException(sprintf('Cache key "%s" contains reserved characters "%s".', $key, $reservedChars));
        }

        return $key;
    }

    /**
     * Validates a cache key according to PSR-6.
     *
     * @param mixed $key The key to validate
     *
     * @throws InvalidArgumentException When $key is not valid
     */
    public static function validateKeyByOverride($key, string $reservedChars = self::RESERVED_CHARACTERS): string
    {
        if (!\is_string($key)) {
            throw new InvalidArgumentException(sprintf('Cache key must be string, "%s" given.', get_debug_type($key)));
        }
        if ('' === $key) {
            throw new InvalidArgumentException('Cache key length must be greater than zero.');
        }
        if ('' !== $reservedChars && false !== strpbrk($key, $reservedChars)) {
            throw new InvalidArgumentException(sprintf('Cache key "%s" contains reserved characters "%s".', $key, $reservedChars));
        }

        return $key;
    }
}

$key = "product:123456:discount";
$iterations = 1000000;

// Testing the original approach
$start_time = microtime(true);
for ($i = 0; $i < $iterations; $i++) {
    CacheItem::validateKey($key, ':');
}
$end_time = microtime(true);
$time_original = $end_time - $start_time;
echo "Original Method Time: " . $time_original . " seconds\n";

// Testing the optimized approach
$start_time = microtime(true);
for ($i = 0; $i < $iterations; $i++) {
    CacheItem::validateKeyByOverride($key, '{}()/\@X');
}
$end_time = microtime(true);
$time_optimized = $end_time - $start_time;
echo "Optimized Method Time: " . $time_optimized . " seconds\n";
```

Here are the results from 5 executions, 1000000 iterations each:

| Original in seconds | Optimized in seconds |
|-----------------|-----------------|
| 1.8678679466248 | 0.62751603126526 |
| 1.4007370471954 | 0.61563205718994 |
| 0.89250898361206 | 0.48939108848572 |
| 0.85805201530457 | 0.41882705688477 |
| 0.86538910865784 | 0.47265005111694 |

Looks like getting rid of the `str_replace` an `str_split` makes the method about 2x faster.

The downside seems relatively small, the logic is moved to the cache adapters. Or we could get rid of the logic completely and just let each adapter define their list of reserved characters..

I am still working on the tests & to ensure that PSR 6 / PSR 16 validation is kept intact.