@upyx 2024-07-24
IMO, it has two **really big** drawbacks:

1. It solves the local issue with Redis by cost of portability with different backends that could not support colons. The Symfony cache is designed as an abstraction that can be implemented by various technologies.
2. It breaks compatibility with PSRs that are gate outside the Symfony world.

I'd rather add a convenient interface to deal with namespaces instead of permitting the namespace separator in keys.