@nicolas-grekas 2025-02-19
Closing in favor of #59813, thanks for giving this a try!